+++ lib/custom_refresh_indicator.dart
@@ -1,12 +1,8 @@
 library custom_refresh_indicator;
 
-import 'dart:ui';
-
-import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
 import 'package:flutter/widgets.dart';
 
-
-part 'src/widget.dart';
+part 'src/helpers/positioned_indicator_container.dart';
+part 'src/custom_refresh_indicator.dart';
 part 'src/definitions.dart';
-part 'src/indicators/simple_indicator_container/simple_indicator_container.dart';