+++ lib/src/custom_refresh_indicator.dart
@@ -11,21 +11,25 @@ class CustomRefreshIndicator extends StatefulWidget {
   final bool leadingGlowVisible;
   final bool trailingGlowVisible;
   final Widget child;
-  final CustomIndicatorBuilder indicatorBuilder;
+  final IndicatorBuilder indicatorBuilder;
+  final ChildTransformBuilder childTransformBuilder;
   final RefreshCallback onRefresh;
 
   CustomRefreshIndicator({
     @required this.child,
     @required this.indicatorBuilder,
     @required this.onRefresh,
+    this.childTransformBuilder,
     this.dragingToIdleDuration = const Duration(milliseconds: 300),
     this.armedToLoadingDuration = const Duration(milliseconds: 200),
     this.loadingToIdleDuration = const Duration(milliseconds: 100),
     this.leadingGlowVisible = false,
     this.trailingGlowVisible = true,
-  })  : assert(child != null),
-        assert(indicatorBuilder != null &&
-            indicatorBuilder is CustomIndicatorBuilder);
+  })  : assert(child != null, 'child argument cannot be null'),
+        assert(
+          indicatorBuilder != null && indicatorBuilder is IndicatorBuilder,
+          'indicatorBuilder argument cannot be null and must be of type CustomIndicatorBuilder',
+        );
 
   @override
   _CustomRefreshIndicatorState createState() => _CustomRefreshIndicatorState();
@@ -33,7 +37,7 @@ class CustomRefreshIndicator extends StatefulWidget {
 
 class _CustomRefreshIndicatorState extends State<CustomRefreshIndicator>
     with TickerProviderStateMixin {
-  /// Whether custom refresh indicator can change [CustomRefreshIndicatorState] from `idle` to `draging`
+  /// Whether custom refresh indicator can change [IndicatorState] from `idle` to `draging`
   bool _canStart = false;
 
   /// Direction in which user is scrolling
@@ -42,26 +46,40 @@ class _CustomRefreshIndicatorState extends State<CustomRefreshIndicator>
   /// The direction in which list scrolls
   AxisDirection _axisDirection;
   double _dragOffset;
-  CustomRefreshIndicatorState _indicatorState;
+  IndicatorState _indicatorState;
 
   AnimationController _animationController;
+  IndicatorData _customRefreshIndicatorData;
+
+  /// Current custom refresh indicator data
+  IndicatorData get data => _customRefreshIndicatorData;
 
   static const double _kPositionLimit = 1.5;
   static const double _kDragContainerExtentPercentage = 0.15;
+  static const double _kInitialValue = 0.0;
 
   @override
   void initState() {
     _dragOffset = 0;
     _canStart = false;
-    _indicatorState = CustomRefreshIndicatorState.idle;
+    _indicatorState = IndicatorState.idle;
     _axisDirection = AxisDirection.down;
     _userScrollingDirection = ScrollDirection.idle;
+
+    _customRefreshIndicatorData = IndicatorData(
+      value: _kInitialValue,
+      direction: _axisDirection,
+      scrollingDirection: _userScrollingDirection,
+      state: _indicatorState,
+    );
 
     _animationController = AnimationController(
       vsync: this,
       upperBound: _kPositionLimit,
-      lowerBound: 0,
-    )..value = 0;
+      lowerBound: 0.0,
+    )
+      ..addListener(_updateCustomRefreshIndicatorData)
+      ..value = _kInitialValue;
 
     super.initState();
   }
@@ -69,10 +87,18 @@ class _CustomRefreshIndicatorState extends State<CustomRefreshIndicator>
   @override
   void dispose() {
     _animationController.dispose();
+    _customRefreshIndicatorData.dispose();
     super.dispose();
   }
 
-  final GlobalKey _key = GlobalKey();
+  void _updateCustomRefreshIndicatorData() {
+    _customRefreshIndicatorData.updateAndNotify(
+      value: _animationController?.value ?? _kInitialValue,
+      direction: _axisDirection,
+      scrollingDirection: _userScrollingDirection,
+      state: _indicatorState,
+    );
+  }
 
   bool _handleGlowNotification(OverscrollIndicatorNotification notification) {
     if (notification.depth != 0) return false;
@@ -86,9 +112,9 @@ class _CustomRefreshIndicatorState extends State<CustomRefreshIndicator>
 
   bool _handleScrollStartNotification(ScrollStartNotification notification) {
     _canStart = notification.metrics.extentBefore == 0 &&
-        _indicatorState == CustomRefreshIndicatorState.idle;
+        _indicatorState == IndicatorState.idle;
 
-    if (_canStart) _indicatorState = CustomRefreshIndicatorState.draging;
+    if (_canStart) _indicatorState = IndicatorState.draging;
 
     _axisDirection = notification.metrics.axisDirection;
     return false;
@@ -96,15 +122,15 @@ class _CustomRefreshIndicatorState extends State<CustomRefreshIndicator>
 
   bool _handleScrollUpdateNotification(ScrollUpdateNotification notification) {
     /// hide when list starts to scroll
-    if (_indicatorState == CustomRefreshIndicatorState.draging ||
-        _indicatorState == CustomRefreshIndicatorState.armed) {
+    if (_indicatorState == IndicatorState.draging ||
+        _indicatorState == IndicatorState.armed) {
       if (notification.metrics.extentBefore > 0.0) {
         _hide();
       } else {
         _dragOffset -= notification.scrollDelta;
         _checkDragOffset(notification.metrics.viewportDimension);
       }
-      if (_indicatorState == CustomRefreshIndicatorState.armed &&
+      if (_indicatorState == IndicatorState.armed &&
           notification.dragDetails == null) {
         _start();
       }
@@ -120,7 +146,7 @@ class _CustomRefreshIndicatorState extends State<CustomRefreshIndicator>
 
   bool _handleScrollEndNotification(ScrollEndNotification notification) {
     if (_animationController.value >= CustomRefreshIndicator.armedFromValue) {
-      if (_indicatorState == CustomRefreshIndicatorState.armed) {
+      if (_indicatorState == IndicatorState.armed) {
         _start();
       }
     } else {
@@ -135,15 +161,15 @@ class _CustomRefreshIndicatorState extends State<CustomRefreshIndicator>
   }
 
   void _checkDragOffset(double containerExtent) {
-    if (_indicatorState == CustomRefreshIndicatorState.hiding ||
-        _indicatorState == CustomRefreshIndicatorState.loading) return;
+    if (_indicatorState == IndicatorState.hiding ||
+        _indicatorState == IndicatorState.loading) return;
     double newValue =
         _dragOffset / (containerExtent * _kDragContainerExtentPercentage);
 
     if (newValue >= CustomRefreshIndicator.armedFromValue) {
-      _indicatorState = CustomRefreshIndicatorState.armed;
+      _indicatorState = IndicatorState.armed;
     } else if (newValue > 0.0) {
-      _indicatorState = CustomRefreshIndicatorState.draging;
+      _indicatorState = IndicatorState.draging;
     }
 
     // triggers indicator update
@@ -169,24 +195,24 @@ class _CustomRefreshIndicatorState extends State<CustomRefreshIndicator>
   void _start() async {
     _dragOffset = 0;
 
-    _indicatorState = CustomRefreshIndicatorState.loading;
+    _indicatorState = IndicatorState.loading;
     await _animationController.animateTo(1.0,
         duration: widget.armedToLoadingDuration);
     await widget.onRefresh();
 
     if (!mounted) return;
-    _indicatorState = CustomRefreshIndicatorState.hiding;
+    _indicatorState = IndicatorState.hiding;
     await _animationController.animateTo(0.0,
         duration: widget.loadingToIdleDuration);
 
     if (!mounted) return;
     setState(() {
-      _indicatorState = CustomRefreshIndicatorState.idle;
+      _indicatorState = IndicatorState.idle;
     });
   }
 
   void _hide() async {
-    _indicatorState = CustomRefreshIndicatorState.hiding;
+    _indicatorState = IndicatorState.hiding;
     _dragOffset = 0;
     _canStart = false;
     await _animationController.animateTo(
@@ -197,34 +223,33 @@ class _CustomRefreshIndicatorState extends State<CustomRefreshIndicator>
 
     if (!mounted) return;
     setState(() {
-      _indicatorState = CustomRefreshIndicatorState.idle;
+      _indicatorState = IndicatorState.idle;
     });
   }
+
+  static final ChildTransformBuilder noChildTransform =
+      (context, child, data) => child;
 
   @override
   Widget build(BuildContext context) => Container(
         child: Stack(
           children: <Widget>[
-            NotificationListener<ScrollNotification>(
-              key: _key,
-              onNotification: _handleScrollNotification,
-              child: NotificationListener<OverscrollIndicatorNotification>(
-                onNotification: _handleGlowNotification,
-                child: widget.child,
-              ),
-            ),
-            AnimatedBuilder(
-              animation: _animationController,
-              builder: (context, snapshot) => widget.indicatorBuilder(
-                context,
-                CustomRefreshIndicatorData(
-                  value: _animationController.value,
-                  direction: _axisDirection,
-                  scrollingDirection: _userScrollingDirection,
-                  state: _indicatorState,
+            (widget.childTransformBuilder ?? noChildTransform).call(
+              context,
+              NotificationListener<ScrollNotification>(
+                onNotification: _handleScrollNotification,
+                child: NotificationListener<OverscrollIndicatorNotification>(
+                  onNotification: _handleGlowNotification,
+                  child: widget.child,
                 ),
               ),
+              data,
             ),
+            if (widget.indicatorBuilder != null)
+              widget.indicatorBuilder(
+                context,
+                _customRefreshIndicatorData,
+              ),
           ],
         ),
       );